Changed Main Activity to allow the recycler view to switch between saved pictures taken using app and default images, also made changes in the adapter to accommodate for this
Changes in PuzzleGridTest include testing a different function to calculate inversions, commented out code trying to implement a timer, and fixing bugs occurring onclick of cells. A bug occurred when clicking (not swipe) cells on the edges of a row when the empty cell index was +/- 1 from it. This allowed the clicked cell to move to this start of the next row where the empty cell was, which is unintended. Resolved this by creating an onclicklistener to check for clicks on valid empty cell neighbours which then calls the swipe move function to process movement instead. Also had a bug in doing this where the old MoveCell function which handled onclick events would be blocked if the cells image tag was same as the empty cell index as I had used the wrong tag index. Resolved by removing this. This was redundant anyway as touches on the empty cell were filtered in the touchListener.