Made a bunch of methods package private instead of private, so they can be accessed in by other classes (activities/adapters).
Updated PuzzleCellView (custom view for cells) to contain code necessary to use it for cells instead of ImageViews, but havent implemented that change yet.

Added timer and move counter functions and graphics. Also added a function to save the lowest time/moves taken for a completed puzzle, and graphics to show them.

Timer works by using a handler to send runnable tasks which updates a graphic every 1s after oncreate is started (up to 1hour).

Best moves/time function simply gets the game data and saves to a private app file if it is better than the saved data (if any).

TODO:
 * Add pause function, must update the timer to function differently first.
 * Allow saving of data for app photo puzzles.
 * UI overlay (fragment?) for pauses or game finishes