Added a UI which opens when the puzzle is solved. It shows the game time and moves and the best for the puzzle(if any).
Changed how to game timer works. It now uses a ScheduledThreadPoolExecutor to post a schedule of delayed runnables every second, with a delay on the first runnable which is set to the remaining time until next tick depending on when the game was paused (0 for start of game). Upon pause, the scheduled tasks are cancelled, and resuming adds a new sequence. This new logic fixes the bug where upon resuming timer, the clock would tick no matter how long till the next tick the game was paused.

Also changed a bunch of logic dealing with pauses and resumes, eg. the timer only starts upon the first move. Think there are no bugs with this logic now, some before included:
 * pausing a game before game timer is even started
 * onResume called after tabbing back to app starts game timer if it had not already started before

 Lastly, added and improved javadocs for methods in PuzzleGridTest