Updated code in MainActivity, ImageRecyclerAdapter, and PuzzleGridTest to fix behaviour of recycler view and choice of puzzle. Also removed ImagePreview class and related xml file and manifest code which was unnecessary.
Fixes:
- User can choose a photo from the recycler view, which will actually load into game now.
- Can make selections in either dataset (photos/defaults) for recycler view, and app will remember each through switching, ie they wont affect each other.
- Taking a photo wont override a selection, but will load in game if no selection (Could change this if wanted).

Note: Recycler 'remembers' selections in either dataset using a separate variable for either which is set to selected images index in its dataset. A boolean is used to track which dataset is being displayed. Images and background for selection are set in holder binding based on the boolean and corresponding tracking variable, as the holders images are shuffled.